requirements.txt
python-dotenv
langchain_openai
langchain_core
langchain_community
streamlit
ipykernel
langchain_cohere
langchain_chroma
pypdf
